backend/requirements.txt
alembic==1.15.1
annotated-types==0.7.0
asgiref==3.8.1
asyncpg==0.30.0
bcrypt==4.3.0
blinker==1.9.0
Brotli==1.1.0
cachelib==0.13.0
click==8.1.8
colorama==0.4.6
dnspython==2.7.0
email_validator==2.2.0
Flask==3.1.0
Flask-Bcrypt==1.0.1
Flask-Caching==2.3.1
Flask-Compress==1.17
flask-cors==5.0.1
Flask-JWT-Extended==4.7.1
Flask-Login==0.6.3
Flask-Migrate==4.1.0
Flask-Principal==0.4.0
Flask-Pydantic==0.12.0
Flask-Security-Too==5.6.1
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
greenlet==3.1.1
idna==3.10
importlib_resources==6.5.2
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.9
MarkupSafe==3.0.2
passlib==1.7.4
psycopg2==2.9.10
psycopg2-binary==2.9.10
pydantic==2.10.6
pydantic_core==2.27.2
PyJWT==2.10.1
python-dotenv==1.0.1
redis==5.2.1
SQLAlchemy==2.0.39
typing_extensions==4.12.2
uuid==1.30
Werkzeug==3.1.3
WTForms==3.2.1
zstandard==0.23.0
